+++ NEWS.md
@@ -8,6 +8,8 @@
 
 1. `DT[, {...; .(A,B)}]` (i.e. when `.()` is the final item of a multi-statement `{...}`) now auto-names the columns `A` and `B` (just like `DT[, .(A,B)]`) rather than `V1` and `V2`, [#2478](https://github.com/Rdatatable/data.table/issues/2478) [#609](https://github.com/Rdatatable/data.table/issues/609). Similarly, `DT[, if (.N>1) .(B), by=A]` now auto-names the column `B` rather than `V1`. Explicit names are unaffected; e.g. `DT[, {... y= ...; .(A=C+y)}, by=...]` named the column `A` before, and still does. Thanks also to @renkun-ken for his go-first strong testing which caught an issue not caught by the test suite or by revdep testing, related to NULL being the last item, [#4061](https://github.com/Rdatatable/data.table/issues/4061).
 
+2. `trunc.cols` in `print.data.table()` option, making it possible to print only the columns that fit in the console without wrapping to new lines. It has a print message that states the number, and the names, of the variables not shown. When `class =  TRUE` in `print.data.table()`, the message also contains the class of the variable.
+
 ## BUG FIXES
 
 ## NOTES